+++ src/main/java/com/esprit/alternance/kaddem/repositories/ContratRepository.java
@@ -3,5 +3,5 @@ package com.esprit.alternance.kaddem.repositories;
 import com.esprit.alternance.kaddem.models.Contrat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContratRepository extends JpaRepository<Contrat,Long> {
+public interface ContratRepository extends JpaRepository<Contrat, Integer> {
 }
