+++ src/main/java/com/esprit/alternance/kaddem/repositories/DetailEquipeRepository.java
@@ -3,5 +3,5 @@ package com.esprit.alternance.kaddem.repositories;
 import com.esprit.alternance.kaddem.models.DetailEquipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetailEquipeRepository extends JpaRepository<DetailEquipe,Long> {
+public interface DetailEquipeRepository extends JpaRepository<DetailEquipe, Integer> {
 }
